moved Serial into prototype package
